move domain code to separate folder
